Move legal page tests into /legal/
